requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
google-cloud-vision
Pillow
python-multipart
